`log` pkg rename `stdout.test.ts` into `exec.test.ts`
